NDIS says at most 64 map registers may be requested in NdisMAllocateMapRegisters, but Atheros driver requests 256! Instead of failing such request, issue warning and continue. The number of registers requested is configurable parameter (MapRegisters) in inf, set by default to 256, so it is possible to change this while installing NDIS driver, but many users have already installed this driver, so for compatibility, tolerate it.

+++ trunk/ndiswrapper/driver/ndis.h
@@ -813,7 +813,7 @@ struct wrap_ndis_device {
 	ULONG packet_filter;
 
 	BOOLEAN use_sg_dma;
-	u8 tx_dma_count;
+	int tx_dma_count;
 	dma_addr_t *tx_dma_addr;
 
 	int hangcheck_interval;
